LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-blocking-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL Block Trusted Type policy creation by event listener. trustedTypes.addEventListener is not a function. (In 'trustedTypes.addEventListener('beforecreatepolicy', event_listener, {once:true})', 'trustedTypes.addEventListener' is undefined)
FAIL Trusted Type policy creation. trustedTypes.addEventListener is not a function. (In 'trustedTypes.addEventListener('beforecreatepolicy', event_listener, {once:true})', 'trustedTypes.addEventListener' is undefined)
FAIL Block only default Trusted Type policy creation. trustedTypes.addEventListener is not a function. (In 'trustedTypes.addEventListener('beforecreatepolicy', event_listener, {once:true})', 'trustedTypes.addEventListener' is undefined)
FAIL Policy creation called before addEventListener function will not reached the listener. trustedTypes.addEventListener is not a function. (In 'trustedTypes.addEventListener('beforecreatepolicy', event_listener, {once:true})', 'trustedTypes.addEventListener' is undefined)
